patch horizon to avoid crashes when creating application credentials
In order to patch reliably, this pins Horizon at a specific commit.
The patch may need updating if the Horizon revision changes (but
hopefully we'll get the patch upstream before that happens).

Because we're pinning at a specific commit we can't use a shallow
clone anymore, so the image size and build time will go up a bit with
this commit.

closes CCI-MOC/ops-issues#4.